requirements.txt
opencv-python==4.11.0.86
numpy==1.26.4
scipy==1.16.2
pillow==10.4.0
degirum==0.18.3
degirum_tools==0.22.2
pycocotools==2.0.10
python-dotenv==1.1.1
